(refactor) move server side into controllers and services plus standardize models
