+++ src/server/models/Card.js
@@ -1,10 +1,10 @@
 import mongoose from '../db';
 
-const CardsSchema = new mongoose.Schema({
+const CardSchema = new mongoose.Schema({
   question: { text: String },
   answer: { text: String, explanation: String },
   deckId: String,
   userId: String,
 }, { timestamps: true });
 
-export default mongoose.model('Cards', CardsSchema);
+export default mongoose.model('Card', CardSchema);
