--- src/server/models/decks.js
@@ -1,7 +1,0 @@
-import mongoose from '../db';
-
-const DecksSchema = new mongoose.Schema({
-  name: String,
-}, { timestamps: true });
-
-export default mongoose.model('Decks', DecksSchema);
