+++ src/server/setupPassport.js
@@ -1,6 +1,6 @@
 import passport from 'passport';
 import { Strategy } from 'passport-local';
-import User from './models/user';
+import User from './models/User';
 
 const oid = '_id';
